suppress warnings
git-svn-id: https://svn.salilab.org/imp/trunk@16768 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/algebra/src/Transformation3D.cpp
@@ -24,8 +24,8 @@ Transformation3D get_transformation_3d(const Transformation2D &t2d) {
   return Transformation3D(R,t);
 }
 Transformation3D get_random_local_transformation(Vector3D origin,
-                       float max_translation,
-                       float max_angle_in_rad) {
+                       double max_translation,
+                       double max_angle_in_rad) {
   algebra::Vector3D translation
     = algebra::get_random_vector_in(algebra::Sphere3D(algebra::Vector3D(0,0,0),
                                                       max_translation));
